Back/requirements.txt
django
djangorestframework
requests
python-whois
dnspython
psycopg2-binary
shodan
python-dotenv
pydantic
python-nmap
openai

# Django>=4.2,<4.3
# djangorestframework>=3.14,<3.15
# requests>=2.28,<2.29
# python-whois>=0.7,<0.8
# dnspython>=2.2,<2.3
# psycopg2-binary>=2.9,<2.10
# shodan>=1.20,<1.21
# python-dotenv>=0.20,<0.22
# pydantic
# python-nmap # Considera usar esta librería para Nmap en lugar de subprocess
